requirements.txt
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
lxml==5.3.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4
numpy==1.26.4
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyMuPDF==1.24.11
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.13.1
six==1.16.0
smart-open==7.0.5
spyder-kernels==2.5.2
stack-data==0.6.3
sympy==1.13.3
tokenizers==0.20.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.16.0
XlsxWriter==3.2.0
